:recycle: Extract target_worker.go from storage.go
